refactor step 1
