+++ src/app/app.module.ts
@@ -28,9 +28,7 @@ import { ElementService } from './service/element.service';
 import { CompoundComponent } from './compound/compound.component';
 import { MatSnackBarModule } from '@angular/material/snack-bar';
 import { TabsComponent } from './tabs/tabs.component';
-import { FlashcardListComponent } from './components/flashcard-list/flashcard-list.component';
 import { FlashcardComponent } from './components/flashcard/flashcard.component';
-import { FlashcardCreateComponent } from './components/flashcard-create/flashcard-create.component';
 import { ProfileComponent } from './profile/profile.component';
 import { EditProfileComponent } from './edit-profile/edit-profile.component';
 import { ValidationModalComponent } from './compound/validation-modal/validation-modal.component';
@@ -47,9 +45,7 @@ import {MatDialogModule, MatDialogRef} from "@angular/material/dialog";
     CompoundComponent,
     TabsComponent,
     UserComponent,
-    FlashcardListComponent,
     FlashcardComponent,
-    FlashcardCreateComponent,
     ProfileComponent,
     EditProfileComponent,
     ValidationModalComponent
