--- src/app/components/flashcard-list/flashcard-list.component.html
@@ -1,8 +1,0 @@
-<div class="container">
-  <div>ICS 499 FlashCards</div>
-  <div class="card-grid">
-      <app-flashcard class="app-flash-card" *ngFor="let flashCard of flashCards; let i = index" [flashcard]="flashCard">
-      </app-flashcard>
-  </div>
-</div>
-
