--- src/app/components/flashcard-list/flashcard-list.component.scss
@@ -1,10 +1,0 @@
-.container {
-  margin: 1rem 2rem;
-}
-
-.card-grid {
-  display: grid;
-  align-items: center;
-  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
-  gap: 1rem;
-}
