--- src/app/components/flashcard-list/flashcard-list.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { FlashcardListComponent } from './flashcard-list.component';
-
-describe('FlashcardListComponent', () => {
-  let component: FlashcardListComponent;
-  let fixture: ComponentFixture<FlashcardListComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ FlashcardListComponent ]
-    })
-    .compileComponents();
-  });
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(FlashcardListComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
