+++ src/app/components/flashcard/flashcard.component.scss
@@ -49,3 +49,14 @@
 .flashcard-option:first-child {
   margin-top: 0;
 }
+
+.container {
+  margin: 1rem 2rem;
+}
+
+.card-grid {
+  display: grid;
+  align-items: center;
+  grid-template-columns: repeat(auto-fill, minmax(200px, 1fr));
+  gap: 1rem;
+}
